src/main/java/cn/yijicai/app/service/impl/UserServiceImpl.java
package cn.yijicai.app.service.impl;

import cn.yijicai.app.mapping.UserMapper;
import cn.yijicai.app.model.User;
import cn.yijicai.app.service.BaseService;
import org.springframework.beans.factory.annotation.Autowired;

import java.util.List;

/**
 * Created by sujian on 2016/4/8.
 */
public class UserServiceImpl implements BaseService<User>{

    @Autowired
    private UserMapper userMapper;

    @Override
    public int deleteByPrimaryKey(Long id) {
        return 0;
    }

    @Override
    public int insert(User record) {
        return 0;
    }

    @Override
    public int insertSelective(User record) {
        return userMapper.insertSelective(record);
    }

    @Override
    public List<User> selectSelective(User record) {
        return null;
    }

    @Override
    public User selectByPrimaryKey(Long id) {
        return userMapper.selectByPrimaryKey(id);
    }

    @Override
    public int updateByPrimaryKeySelective(User record) {
        return 0;
    }

    @Override
    public int updateByPrimaryKey(User record) {
        return 0;
    }
}
